% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_functions.R
\name{compute_kmeans}
\alias{compute_kmeans}
\title{Perform K-means Clustering}
\usage{
compute_kmeans(data, k)
}
\arguments{
\item{data}{A data frame with at least two numeric columns.}

\item{k}{The number of clusters.}
}
\value{
A k-means model.
}
\description{
This function applies K-means clustering to a given dataset.
}
