% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_functions.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Plot K-means Clustering Results}
\usage{
plot_kmeans(data, kmeans_model)
}
\arguments{
\item{data}{A data frame with at least two numeric columns.}

\item{kmeans_model}{A k-means model object from \code{compute_kmeans()}.}
}
\value{
A ggplot2 object.
}
\description{
This function plots K-means clustering results using ggplot2.
}
